Allow all inspections to be disabled
I have some auto-generated code (from scalaxb for example) that fails... a lot of checks.  Given that a human isn't responsible for making sure it works, I disable all 'linters' on those projects which were entirely made by a script / tool.  I tried -Xplugin-disable as I do other ones but then I get odd errors relating to settings or something... this seems like the an easy solution.  Guessed a good way to do this but if you have a better way feel free to ignore this.

Also I just made the changes in GitHub so I apologize if there are any syntax errors.