ComparingUnrelatedTypes: expand to cover parameterized types
Previously this check would fail with an exception on parameterized types because
the .asClass method will throw on parameterized types.

This patch fixes that and additionally expands the checks to cover parameterized types.
Note that the the check will give false positives on comparisons like warning 7.
I'm not familiar enough with the scala reflection api to figure out how to deal with
that case at the moment.