Fix ComparingUnrelatedTypes false-positives for comparing to integral type literals.
The ComparingUnrelatedTypes rule would often trigger false positives in comparisons
against integral type literals. For example,

```
val x = 1
c == 100L
```

would trigger a false positive even though this is a safe comparision because the long value `100L`
fits within an `Int` type.

To avoid such false-positive, this patch extends the ComparingUnrelatedTypes rule to add
special-case checks for the comparision of a non-constant integral value to a literal integral
value. In such cases, we check whether the literal is a valid value of the other value's type.
Thus `(c: Char) == 97` is considered valid but `(c: Char) == 128000` is not.