add Map.getAndGetOrElse inspection
If you need to lookup a key in a Map and then return a default value
if the key does not exist in the Map such as:

```
someMap.get(key).getOrElse(defaultValue)
```

This inspection will propose this simplified code instead:

```
someMap.getOrElse(key, defaultValue)
```

which does the same thing.

Inspired by the Intellij inspection that does the same.