Fix spurious UnreachableCatch when some catch cases are conditional
In a try...catch that looks like this:

```
{
  case e: SomeException if condition(e) =>
  case e: SomeException if otherCondition(e) =>
}
```

it would incorrectly report the second case as unreachable, since it didn't
actually examine the `if ...` guards attached to each case.

This change makes it so that if a case is conditional (has a non-empty guard
attached to it), then it's not considered to be redundant with subsequent cases
against the same type, though it is still checked against
previous (unconditional) cases that might have had the same type.

Also fix the formatting of the message, which was missing a space.