Fix line numbers and snippets for VarCouldBeVal
Previously, the VarCouldBeVal inspection attributed any errors not to the line
where the "var" was defined, but to the line where the enclosing def
started. (The snippet also showed, potentially, all of this enclosing def.)

Instead, we now remember the tree (of type ValDef) that defines each var, and
if it turns out that it's unwritten, we output the line number and snippet
corresponding to that tree.