Publish with full Scala version
As a rule of thumb, a scalac plugin always should be published using the full
scala version in its name, since scalac's internal APIs can change without notice.

This change introduces the sbt-travisci plugin which makes it easier to
maintain a full set of versions, cross compiled and run through the CI.

Fixes #248.